refactor tools into pkg
